beta

+++ Pangu.spritebuilder/Source/Gameplay.m
@@ -115,7 +115,7 @@ static Gameplay* s_currentGame;
     NSNumber* indexNumber = [NSNumber numberWithInteger:index];
     NSDictionary* actor = [_currentScript.script dictFrom:@[@"actors", indexNumber]] ;
     NSArray* roles = [actor arrayFrom:@[@"roles"]];
-    NSInteger delay = [actor doubleFrom:@[@"delay"]];
+    CGFloat delay = [actor doubleFrom:@[@"delay"]];
     if (delay < MIN_UNIT) {
         delay = [_currentScript.script doubleFrom:@[@"delay"]];
     }
